Refactor classes responsibility
