+++ Golem/Yagna/YagnaService.cs
@@ -23,38 +23,6 @@ namespace Golem.Yagna
         public PaymentDriver PaymentDriver { get; set; } = PaymentDriver.ERC20;
     }
 
-
-
-    //[JsonObject(MemberSerialization.OptIn)]
-    public class IdInfo
-    {
-        // [JsonPropertyName("alias")]
-        public string? Alias { get; set; }
-
-        // [JsonPropertyName("default")]
-        public bool IsDefault { get; set; }
-
-        // [JsonPropertyName("locked")]
-        public bool IsLocked { get; set; }
-
-        // [JsonPropertyName("nodeId")]
-        public string NodeId { get; set; }
-
-        public IdInfo(bool _isDefault, bool _isLocked, string? _alias, string _nodeId)
-        {
-            IsDefault = _isDefault;
-            IsLocked = _isLocked;
-            Alias = _alias;
-            NodeId = _nodeId;
-        }
-
-        [JsonConstructor]
-        public IdInfo()
-        {
-            NodeId = "";
-        }
-    }
-
     public class YagnaService
     {
         private readonly string _yaExePath;
@@ -321,195 +289,5 @@ namespace Golem.Yagna
         {
             _logger.LogInformation($"{e.Data}");
         }
-    }
-
-    public class KeyInfo
-    {
-        public string Name { get; set; }
-
-        public string? Key { get; set; }
-
-        public string Id { get; set; }
-
-        public string? Role { get; set; }
-
-        public DateTime? Created { get; set; }
-
-        [JsonConstructor]
-        public KeyInfo()
-        {
-            Name = "";
-            Id = "";
-        }
-    }
-
-    public class MeInfo
-    {
-        public string? Name { get; set; }
-        public string? Identity { get; set; }
-
-        public string? Role { get; set; }
-
-        [JsonConstructor]
-        public MeInfo()
-        {
-        }
-    }
-
-    public class AppKeyService
-    {
-        private readonly YagnaService _yagnaService;
-        private readonly string? _id;
-
-        internal AppKeyService(YagnaService yagnaService, string? id)
-        {
-            this._yagnaService = yagnaService;
-            this._id = id;
-        }
-
-        private string[] prepareArgs(params string[] arguments)
-        {
-            var execArgs = new List<string>(3 + arguments.Length);
-            execArgs.Add("--json");
-            if (_id != null)
-            {
-                execArgs.Add("--id");
-                execArgs.Add(_id);
-            }
-            execArgs.Add("app-key");
-            execArgs.AddRange(arguments);
-
-            return execArgs.ToArray();
-        }
-
-        private T? Exec<T>(params string[] arguments) where T : class
-        {
-            return _yagnaService.Exec<T>(prepareArgs(arguments));
-        }
-
-        private async Task<T?> ExecAsync<T>(params string[] arguments) where T : class
-        {
-            return await _yagnaService.ExecAsync<T>(prepareArgs(arguments));
-        }
-
-        public string? Create(string name)
-        {
-            return Exec<string>("create", name);
-        }
-        public async Task<string?> CreateAsync(string name)
-        {
-            return await ExecAsync<string>("create", name);
-        }
-
-        public void Drop(string name)
-        {
-            var opt = Exec<string>("drop", name);
-        }
-
-        public KeyInfo? Get(string name)
-        {
-            var keys = List();
-            if (keys is not null && keys.Count > 0)
-            {
-                if (keys.Count > 1)
-                {
-                    return keys.Where(x => x.Name.Equals(name)).FirstOrDefault();
-                }
-                else if (keys.Count == 1)
-                {
-                    return keys.First();
-                }
-            }
-            return null;
-        }
-
-        public List<KeyInfo> List()
-        {
-            var output = new List<KeyInfo>();
-            int tries = 0;
-            while (output.Count == 0)
-            {
-                List<KeyInfo>? o;
-                try
-                {
-                    o = Exec<List<KeyInfo>>("list");
-                }
-                catch (Exception e)
-                {
-                    Console.Error.WriteLine(e.ToString());
-                    o = null;
-                }
-                if (o != null)
-                    output.AddRange(o);
-                else
-                    Thread.Sleep(1000);
-                if (++tries == 10)
-                {
-                    throw new Exception("Failed to obtain key list from yagna service");
-                }
-            }
-            return output;
-        }
-
-        public async Task<List<KeyInfo>> ListAsync()
-        {
-            var output = await ExecAsync<List<KeyInfo>>("list");
-            return output ?? new List<KeyInfo>();
-        }
-
-
-    }
-
-    public class IdService
-    {
-        readonly YagnaService _yagna;
-
-        internal IdService(YagnaService yagna)
-        {
-            _yagna = yagna;
-        }
-
-        public List<IdInfo> List()
-        {
-            var table = _yagna.Exec<List<IdInfo>>("--json", "id", "list");
-            return table ?? new List<IdInfo>();
-        }
-    }
-
-
-    public class PaymentService
-    {
-        readonly YagnaService _yagna;
-
-        internal PaymentService(YagnaService yagna)
-        {
-            _yagna = yagna;
-        }
-
-        public void Init(string account)
-        {
-            var yagnaOptions = _yagna.StartupOptions();
-            _yagna.ExecToText("payment", "init", "--receiver", "--network", yagnaOptions.Network.Id, "--driver", yagnaOptions.PaymentDriver.Id, "--account", account);
-        }
-
-        public async Task<PaymentStatus?> Status(Network network, string driver, string account)
-        {
-            return await _yagna.ExecAsync<PaymentStatus>("--json", "payment", "status", "--network", network.Id, "--driver", driver, "--account", account);
-        }
-
-        public async Task<string?> ExitTo(Network network, string driver, string account, string? destination)
-        {
-            if (destination == null)
-            {
-                return null;
-            }
-            return await _yagna.ExecToTextAsync("--json", "payment", "exit", "--network", network.Id, "--driver", driver, "--account", account, "--to-address", destination);
-        }
-
-        public async Task<ActivityStatus?> ActivityStatus()
-        {
-            return await _yagna.ExecAsync<ActivityStatus>("--json", "activity", "status");
-        }
-
     }
 }
